Приведение всех id к типу Long

+++ src/main/java/gui/model/Conversion.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class Conversion {
 
-    private Integer idConversion;
+    private Long idConversion;
     private String nameConversion;
     private GroupConversion idGroup;
 }
